Backend/requirements.txt
Flask==3.0.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.7.1
flask-cors==5.0.0
pymongo==4.10.1
google-generativeai==0.8.3
python-dotenv==0.21.0
